requirements.txt
numpy==2.0.1
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
matplotlib==3.10.1
scikit-learn
scikit-image
tqdm
h5py
pillow==11.1.0
opencv-python
pandas==2.2.3
timm==1.0.15

--extra-index-url https://download.pytorch.org/whl/cu121